requirements.txt
deprecation>=2.1.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
pennprov==2.2.4
python-dotenv>=1.0.0
ipykernel>=6.0.0